rename: 회원 관련 디렉토리명 수정에 따른 변경
