+++ src/main/java/io/woori/account/wooriaccount/dto/customer/LoginRequestDTO.java
@@ -1,11 +1,15 @@
-package io.woori.account.wooriaccount.dto.user;
+package io.woori.account.wooriaccount.dto.customer;
 
 import lombok.*;
+
+import java.io.Serializable;
 
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class LoginRequestDTO {
+public class LoginRequestDTO implements Serializable {
+
     private String email; // 이메일 (로그인 id)
     private String pwd; // 로그인 비번
+
 }
